Demo/bookmanager/requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
google-generativeai==0.3.2 